// src/shared/generated/GraphicsStatusRecord.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { GraphicsSettingsRecord } from "./GraphicsSettingsRecord";

export type GraphicsStatusRecord = { supported: boolean, sequence: number, initialized: boolean, hardwareAcceleration: boolean | null, appliedSettings: GraphicsSettingsRecord | null, features: { [key in string]: string }, devices: Array<{ [key in string]: string }>, driver: { [key in string]: string }, versions: { [key in string]: string }, problems: Array<string>, complete: boolean, error: string | null, };